+++ src/env.rs
@@ -128,7 +128,6 @@ impl<'de> serde::Deserializer<'de> for Deserializer {
     }
 }
 
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -165,4 +164,106 @@ mod tests {
         assert_eq!(de("3.1415"), Ok(3.1415f32));
         assert_eq!(de("-123.456"), Ok(-123.456f64));
     }
+}
+
+/// This module contains helper methods to simplify configuration via environment variables
+pub mod env_utils {
+    use std::str::FromStr;
+
+    /// Helper function to parse any type that implements [`std::str::FromStr`]
+    /// into a collection that implements [`std::iter::FromIterator`], spliting original string by
+    /// const char separator.
+    ///
+    /// # To Vec
+    /// ```
+    /// use crate::confique::Config;
+    /// #[derive(Debug, confique::Config)]
+    /// struct Conf {
+    ///     #[config(
+    ///         env = "PORTS",
+    ///         parse_env = confique::env_utils::to_collection_by_char_separator::<',', _, _>
+    ///     )]
+    ///     ports: Vec<u16>
+    /// }
+    ///
+    /// std::env::set_var("PORTS", "8080,8000,8888");
+    /// println!("{:?}", Conf::builder().env().load().unwrap())
+    /// ```
+    ///
+    /// # To HashSet
+    /// ```
+    /// use crate::confique::Config;
+    /// #[derive(Debug, confique::Config)]
+    /// struct Conf {
+    ///     #[config(
+    ///         env = "PATHS",
+    ///         parse_env = confique::env_utils::to_collection_by_char_separator::<';', _, _>
+    ///     )]
+    ///     paths: std::collections::HashSet<std::path::PathBuf>,
+    /// }
+    ///
+    /// std::env::set_var("PATHS", "/bin;/user/bin;/home/user/.cargo/bin");
+    /// println!("{:?}", Conf::builder().env().load().unwrap())
+    /// ```
+    pub fn to_collection_by_char_separator<
+        const SEPARATOR: char,
+        T: FromStr,
+        C: FromIterator<Result<T, <T as FromStr>::Err>>,
+    >(
+        input: &str,
+    ) -> C {
+        input.split(SEPARATOR.to_owned()).map(T::from_str).collect()
+    }
+
+    macro_rules! specify_fn_wrapper {
+        ($symbol_name:ident, $symbol:tt) => {
+            ::paste::paste! {
+                /// Helper function to parse any type that implements [`std::str::FromStr`]
+                /// into a collection that implements [`std::iter::FromIterator`], spliting original string by
+                #[doc = stringify!($symbol_name)]
+                /// .
+                ///
+                /// # To Vec
+                /// ```
+                /// use crate::confique::Config;
+                /// #[derive(Debug, confique::Config)]
+                /// struct Conf {
+                ///     #[config(
+                ///         env = "PORTS",
+                #[doc = concat!("         parse_env = confique::env_utils::", stringify!([<to_collection_by_ $symbol_name>],))]
+                ///     )]
+                ///     ports: Vec<u16>
+                /// }
+                ///
+                #[doc = concat!("std::env::set_var(\"PORTS\", \"8080", $symbol, "8000", $symbol, "8888", "\");")]
+                /// println!("{:#?}", Conf::builder().env().load().unwrap())
+                /// ```
+                ///
+                /// # To HashSet
+                /// ```
+                /// use crate::confique::Config;
+                /// #[derive(Debug, confique::Config)]
+                /// struct Conf {
+                ///     #[config(
+                ///         env = "PATHS",
+                #[doc = concat!("         parse_env = confique::env_utils::", stringify!([<to_collection_by_ $symbol_name>],))]
+                ///     )]
+                ///     paths: std::collections::HashSet<std::path::PathBuf>,
+                /// }
+                ///
+                #[doc = concat!("std::env::set_var(\"PATHS\", \"/bin", $symbol, "/user/bin", $symbol, "/home/user/.cargo/bin", "\");")]
+                /// println!("{:#?}", Conf::builder().env().load().unwrap())
+                /// ```
+                pub fn [<to_collection_by_ $symbol_name>]<T: FromStr, C: FromIterator<Result<T, <T as FromStr>::Err>>>(
+                    input: &str,
+                ) -> C {
+                    to_collection_by_char_separator::<$symbol, _, _>(input)
+                }
+            }
+        }
+    }
+
+    specify_fn_wrapper!(comma, ',');
+    specify_fn_wrapper!(semicolon, ';');
+    specify_fn_wrapper!(space, ' ');
 }