+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 
 
 ## [Unreleased]
-
+### Added
+- `parse_env` attribute which is using for parse env variable in arbitrary type
 
 ## [0.2.0] - 2022-10-21
 ### Added